[Maintenance] Move invoice identifier generator to generator namespace
