+++ src/Generator/InvoiceIdentifierGenerator.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Sylius\InvoicingPlugin\Generator;
+
+interface InvoiceIdentifierGenerator
+{
+    public function __invoke(string $orderNumber): string;
+}
